Minor Changes due to refactoring (renaming packages)
OPEN - task 54: Create Category CRUD packages 
http://github.com/DevOpsDistilled/OpERP/issues/issue/54
