+++ OpERP/src/main/java/devopsdistilled/operp/client/items/models/observers/BrandModelObserver.java
@@ -2,10 +2,10 @@ package devopsdistilled.operp.client.items.models.observers;
 
 import java.util.List;
 
-import devopsdistilled.operp.client.abstracts.EntityObserver;
+import devopsdistilled.operp.client.abstracts.EntityModelObserver;
 import devopsdistilled.operp.server.data.entity.items.Brand;
 
-public interface BrandModelObserver extends EntityObserver {
+public interface BrandModelObserver extends EntityModelObserver {
 
 	public void updateBrands(List<Brand> brands);
