+++ OpERP/src/main/java/devopsdistilled/operp/client/items/models/observers/ItemModelObserver.java
@@ -2,10 +2,10 @@ package devopsdistilled.operp.client.items.models.observers;
 
 import java.util.List;
 
-import devopsdistilled.operp.client.abstracts.EntityObserver;
+import devopsdistilled.operp.client.abstracts.EntityModelObserver;
 import devopsdistilled.operp.server.data.entity.items.Item;
 
-public interface ItemModelObserver extends EntityObserver {
+public interface ItemModelObserver extends EntityModelObserver {
 
 	public void updateItems(List<Item> items);
