+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/controllers/impl/ListItemPaneControllerImpl.java
@@ -5,7 +5,7 @@ import javax.inject.Inject;
 import devopsdistilled.operp.client.items.models.ItemModel;
 import devopsdistilled.operp.client.items.panes.ListItemPane;
 import devopsdistilled.operp.client.items.panes.controllers.ListItemPaneController;
-import devopsdistilled.operp.client.items.panes.model.ListItemPaneModel;
+import devopsdistilled.operp.client.items.panes.models.ListItemPaneModel;
 
 public class ListItemPaneControllerImpl implements ListItemPaneController {
