+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/CreateProductPaneModel.java
@@ -1,7 +1,7 @@
-package devopsdistilled.operp.client.items.panes.model;
+package devopsdistilled.operp.client.items.panes.models;
 
 import devopsdistilled.operp.client.abstracts.PaneModel;
-import devopsdistilled.operp.client.items.panes.model.observers.CreateProductPaneModelObserver;
+import devopsdistilled.operp.client.items.panes.models.observers.CreateProductPaneModelObserver;
 
 public interface CreateProductPaneModel extends
 		PaneModel<CreateProductPaneModelObserver> {
