+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/ListItemPaneModel.java
@@ -1,4 +1,4 @@
-package devopsdistilled.operp.client.items.panes.model;
+package devopsdistilled.operp.client.items.panes.models;
 
 import devopsdistilled.operp.client.abstracts.PaneModel;
 import devopsdistilled.operp.client.items.models.observers.ListItemPaneModelObserver;
