+++ OpERP/src/main/java/devopsdistilled/operp/client/items/panes/models/impl/ListItemPaneModelImpl.java
@@ -1,8 +1,8 @@
-package devopsdistilled.operp.client.items.panes.model.impl;
+package devopsdistilled.operp.client.items.panes.models.impl;
 
 import devopsdistilled.operp.client.abstracts.AbstractPaneModel;
 import devopsdistilled.operp.client.items.models.observers.ListItemPaneModelObserver;
-import devopsdistilled.operp.client.items.panes.model.ListItemPaneModel;
+import devopsdistilled.operp.client.items.panes.models.ListItemPaneModel;
 
 public class ListItemPaneModelImpl extends
 		AbstractPaneModel<ListItemPaneModelObserver> implements
